# go-dianping/api/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Go 黑马点评 API",
    "version": "1.0.0",
    "description": "保留Java业务Result格式。教学写接口需登录但没有商家RBAC；公开部署前应补角色权限。金额为分；ID为int64，秒杀响应X-Order-ID为精确字符串。"
  },
  "servers": [
    {
      "url": "http://localhost:8081"
    },
    {
      "url": "http://localhost:8080/api"
    }
  ],
  "paths": {
    "/user/code": {
      "post": {
        "summary": "发送验证码",
        "operationId": "post__user_code",
        "tags": [
          "user"
        ],
        "security": [],
        "description": "开发短信写API日志，不在响应中泄露。关闭dev_code_log时明确返回503，需接短信商。\nRedis: String login:code:手机号 2m；login:throttle:手机号 60s\nMySQL: 无",
        "parameters": [
          {
            "name": "phone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "11位中国大陆手机号",
            "example": "13900000001"
          }
        ],
        "responses": {
          "200": {
            "description": "业务结果；检查success字段",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "success"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true
                    }
                  },
                  "failure": {
                    "value": {
                      "success": false,
                      "errorMsg": "验证码发送频繁，请稍后重试"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "缺少或失效Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "系统故障",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖或短信服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "summary": "验证码/密码登录",
        "operationId": "post__user_login",
        "tags": [
          "user"
        ],
        "security": [],
        "description": "裸Token返回data；password分支只验证bcrypt，不接受原未启用的MD5工具。\nRedis: 消费验证码；限制5次尝试；写login:token:Token Hash，TTL30m\nMySQL: 按唯一手机号查询；首次验证码登录创建tb_user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "业务结果；检查success字段",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "success"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "data": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": "64位随机十六进制Token"
                    }
                  },
                  "failure": {
                    "value": {
                      "success": false,
                      "errorMsg": "验证码无效或尝试次数过多"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "缺少或失效Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "系统故障",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖或短信服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              },
              "example": {
                "phone": "13900000001",
                "code": "从日志或Redis读取"
              }
            }
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "summary": "注销当前会话",
        "operationId": "post__user_logout",
        "tags": [
          "user"
        ],
        "security": [
          {
            "token": []
          }
        ],
        "description": "修复Java仅清ThreadLocal却未删除登录态的问题。\nRedis: DEL当前Token Hash\nMySQL: 无",
        "parameters": [],
        "responses": {
          "200": {
            "description": "业务结果；检查success字段",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "success"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true
                    }
                  },
                  "failure": {
                    "value": {
                      "success": false,
                      "errorMsg": "请求参数无效"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "缺少或失效Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "系统故障",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖或短信服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/refresh": {
      "post": {
        "summary": "主动续期",
        "operationId": "post__user_refresh",
        "tags": [
          "user"
        ],
        "security": [
          {
            "token": []
          }
        ],
        "description": "新增；平常任意携带有效Token的请求也自动滑动续期。\nRedis: 中间件把当前Token TTL刷新至30m\nMySQL: 无",
        "parameters": [],
        "responses": {
          "200": {
            "description": "业务结果；检查success字段",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "success"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "data": {
                          "type": "object",
                          "properties": {
                            "ttlSeconds": {
                              "type": "integer",
                              "format": "int64"
                            }
                          }
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "ttlSeconds": 1800
                      }
                    }
                  },
                  "failure": {
                    "value": {
                      "success": false,
                      "errorMsg": "请求参数无效"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "缺少或失效Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "系统故障",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖或短信服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/me": {
      "get": {
        "summary": "当前用户",
        "operationId": "get__user_me",
        "tags": [
          "user"
        ],
        "security": [
          {
            "token": []
          }
        ],
        "description": "\nRedis: 读取并刷新当前Token Hash\nMySQL: 无",
        "parameters": [],
        "responses": {
          "200": {
            "description": "业务结果；检查success字段",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "success"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "data": {
                          "$ref": "#/components/schemas/UserDTO"
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "id": 1,
                        "nickName": "学习用户一",
                        "icon": ""
                      }
                    }
                  },
                  "failure": {
                    "value": {
                      "success": false,
                      "errorMsg": "请求参数无效"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "缺少或失效Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "系统故障",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖或短信服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "summary": "查询用户公开信息",
        "operationId": "get__user__id_",
        "tags": [
          "user"
        ],
        "security": [
          {
            "token": []
          }
        ],
        "description": "不存在返回success:true无data，不返回phone/password。\nRedis: 仅会话续期\nMySQL: SELECT tb_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "description": "目标资源ID，正整数",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "业务结果；检查success字段",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "success"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "data": {
                          "$ref": "#/components/schemas/UserDTO"
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "id": 1,
                        "nickName": "学习用户一",
                        "icon": ""
                      }
                    }
                  },
                  "failure": {
                    "value": {
                      "success": false,
                      "errorMsg": "请求参数无效"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "缺少或失效Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "系统故障",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖或短信服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/info/{id}": {
      "get": {
        "summary": "用户资料",
        "operationId": "get__user_info__id_",
        "tags": [
          "user"
        ],
        "security": [
          {
            "token": []
          }
        ],
        "description": "无资料时无data；不返回创建/更新时间。\nRedis: 仅会话续期\nMySQL: SELECT tb_user_info",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "description": "目标资源ID，正整数",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "业务结果；检查success字段",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "success"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "data": {
                          "$ref": "#/components/schemas/UserInfo"
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "userId": 1,
                        "city": "杭州",
                        "gender": 0,
                        "level": 0
                      }
                    }
                  },
                  "failure": {
                    "value": {
                      "success": false,
                      "errorMsg": "请求参数无效"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "缺少或失效Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "系统故障",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖或短信服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "直接填Token；也支持Bearer Token"
      }
    },
    "schemas": {
...
      "UserInfo": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "用户 ID"
          },
          "city": {
            "type": "string",
            "description": "城市"
          },
          "introduce": {
            "type": "string",
            "description": "简介"
          },
          "fans": {
            "type": "integer",
            "format": "int64",
            "description": "粉丝数量（原用户资料字段，不随关注实时统计）"
          },
          "followee": {
            "type": "integer",
            "format": "int64",
            "description": "关注数（原资料字段）"
          },
          "gender": {
            "type": "integer",
            "format": "int64",
            "description": "0男/1女（沿用原数据）"
          },
          "birthday": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "生日；当前输出RFC3339日期时间"
          },
          "credits": {
            "type": "integer",
            "format": "int64",
            "description": "积分"
          },
          "level": {
            "type": "integer",
            "format": "int64",
            "description": "会员等级0–9"
          }
        }
      },
...
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键；int64。订单 ID 可能超过 JavaScript 精确整数范围。"
          },
          "nickName": {
            "type": "string",
            "description": "昵称"
          },
          "icon": {
            "type": "string",
            "description": "头像/图标路径"
          }
        }
      },
      "LoginInput": {
        "type": "object",
        "required": [
          "phone"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^1[3-9][0-9]{9}$"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "description": "验证码优先于password；仅可成功使用一次"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72,
            "description": "新增bcrypt密码登录；先用验证码登录设置密码"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "success",
          "errorMsg"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "errorMsg": {
            "type": "string"
          }
        }
      }
    }
  }
}